separate navigation logic in post viewer to hook
